droply_file_api/requirements.txt
flask
python-dotenv
psycopg2-binary
flasgger
flask-cors